初始版负载均衡
一版负载均衡
优化了单例，只保留get_ins接口

+++ lib/number.py
@@ -32,4 +32,4 @@ def consume_cpu_time(consume_time):
 # 上面的改进版
 def consume_cpu_time_1(consume_time):
     # 差不多6500000 ~= 1秒
-    positive_integer_sum(consume_time * 6500000)
+    positive_integer_sum(int(consume_time * 6500000))
